+++ sklearn/ensemble/forest.py
@@ -42,6 +42,7 @@ from ..feature_selection.selector_mixin import SelectorMixin
 from ..tree import DecisionTreeClassifier, DecisionTreeRegressor, \
                    ExtraTreeClassifier, ExtraTreeRegressor
 from ..utils import check_random_state
+from ..metrics import r2_score
 
 from .base import BaseEnsemble
 
@@ -71,6 +72,7 @@ def _parallel_build_trees(n_trees, forest, X, y,
             indices = random_state.randint(0, n_samples, n_samples)
             tree.fit(X[indices], y[indices],
                      sample_mask=sample_mask, X_argsorted=X_argsorted)
+            tree.indices_ = indices
 
         else:
             tree.fit(X, y,
@@ -137,6 +139,7 @@ class BaseForest(BaseEnsemble, SelectorMixin):
                        estimator_params=[],
                        bootstrap=False,
                        compute_importances=False,
+                       oob_score=False,
                        n_jobs=1,
                        random_state=None):
         super(BaseForest, self).__init__(
@@ -146,6 +149,7 @@ class BaseForest(BaseEnsemble, SelectorMixin):
 
         self.bootstrap = bootstrap
         self.compute_importances = compute_importances
+        self.oob_score = oob_score
         self.n_jobs = n_jobs
         self.random_state = check_random_state(random_state)
 
@@ -177,6 +181,10 @@ class BaseForest(BaseEnsemble, SelectorMixin):
             X_argsorted = None
 
         else:
+            if self.oob_score:
+                raise ValueError("Out of bag estimation only available"
+                        " if bootstrap=True")
+
             sample_mask = np.ones((X.shape[0],), dtype=np.bool)
             X_argsorted = np.asfortranarray(
                 np.argsort(X.T, axis=1).astype(np.int32).T)
@@ -204,6 +212,33 @@ class BaseForest(BaseEnsemble, SelectorMixin):
         # Reduce
         self.estimators_ = [tree for tree in itertools.chain(*all_trees)]
 
+        # Calculate out of bag predictions and score
+        if self.oob_score:
+            if isinstance(self, ClassifierMixin):
+                predictions = np.zeros((X.shape[0], self.n_classes_))
+                for estimator in self.estimators_:
+                    mask = np.ones(X.shape[0], dtype=np.bool)
+                    mask[estimator.indices_] = False
+                    predictions[mask, :] += estimator.predict_proba(X[mask, :])
+
+                self.oob_decision_function_ = (predictions
+                        / predictions.sum(axis=1)[:, np.newaxis])
+                self.oob_score_ = np.mean(y == np.argmax(predictions, axis=1))
+
+            else:
+                # Regression:
+                predictions = np.zeros(X.shape[0])
+                n_predictions = np.zeros(X.shape[0])
+                for estimator in self.estimators_:
+                    mask = np.ones(X.shape[0], dtype=np.bool)
+                    mask[estimator.indices_] = False
+                    predictions[mask] += estimator.predict(X[mask, :])
+                    n_predictions[mask] += 1
+                predictions /= n_predictions
+
+                self.oob_prediction_ = predictions
+                self.oob_score_ = r2_score(y, predictions)
+
         # Sum the importances
         if self.compute_importances:
             self.feature_importances_ = \
@@ -224,6 +259,7 @@ class ForestClassifier(BaseForest, ClassifierMixin):
                        estimator_params=[],
                        bootstrap=False,
                        compute_importances=False,
+                       oob_score=False,
                        n_jobs=1,
                        random_state=None):
         super(ForestClassifier, self).__init__(
@@ -232,6 +268,7 @@ class ForestClassifier(BaseForest, ClassifierMixin):
             estimator_params=estimator_params,
             bootstrap=bootstrap,
             compute_importances=compute_importances,
+            oob_score=oob_score,
             n_jobs=n_jobs,
             random_state=random_state)
 
@@ -320,6 +357,7 @@ class ForestRegressor(BaseForest, RegressorMixin):
                        estimator_params=[],
                        bootstrap=False,
                        compute_importances=False,
+                       oob_score=False,
                        n_jobs=1,
                        random_state=None):
         super(ForestRegressor, self).__init__(
@@ -328,6 +366,7 @@ class ForestRegressor(BaseForest, RegressorMixin):
             estimator_params=estimator_params,
             bootstrap=bootstrap,
             compute_importances=compute_importances,
+            oob_score=oob_score,
             n_jobs=n_jobs,
             random_state=random_state)
 
@@ -421,6 +460,10 @@ class RandomForestClassifier(ForestClassifier):
         Whether feature importances are computed and stored into the
         ``feature_importances_`` attribute when calling fit.
 
+    oob_score : bool
+        Whether to use out-of-bag samples to estimate
+        the generalization error.
+
     n_jobs : integer, optional (default=1)
         The number of jobs to run in parallel. If -1, then the number of jobs
         is set to the number of cores.
@@ -433,8 +476,15 @@ class RandomForestClassifier(ForestClassifier):
 
     Attributes
     ----------
-    `feature_importances_` : array of shape = [n_features]
-        The feature mportances (the higher, the more important the feature).
+    `feature_importances_` : array, shape = [n_features]
+        The feature importances (the higher, the more important the feature).
+
+    `oob_score_` : float
+        Score of the training dataset obtained using an out-of-bag estimate.
+
+    `oob_decision_function_` : array, shape = [n_samples, n_classes]
+        Decision function computed with out-of-bag estimate on the training set.
+
 
     Notes
     -----
@@ -454,6 +504,7 @@ class RandomForestClassifier(ForestClassifier):
                        max_features="auto",
                        bootstrap=True,
                        compute_importances=False,
+                       oob_score=False,
                        n_jobs=1,
                        random_state=None):
         super(RandomForestClassifier, self).__init__(
@@ -463,6 +514,7 @@ class RandomForestClassifier(ForestClassifier):
                               "min_density", "max_features", "random_state"),
             bootstrap=bootstrap,
             compute_importances=compute_importances,
+            oob_score=oob_score,
             n_jobs=n_jobs,
             random_state=random_state)
 
@@ -529,6 +581,10 @@ class RandomForestRegressor(ForestRegressor):
         Whether feature importances are computed and stored into the
         ``feature_importances_`` attribute when calling fit.
 
+    oob_score : bool
+        whether to use out-of-bag samples to estimate
+        the generalization error.
+
     n_jobs : integer, optional (default=1)
         The number of jobs to run in parallel. If -1, then the number of jobs
         is set to the number of cores.
@@ -543,6 +599,14 @@ class RandomForestRegressor(ForestRegressor):
     ----------
     `feature_importances_` : array of shape = [n_features]
         The feature mportances (the higher, the more important the feature).
+
+    `oob_score_` : float
+        Score of the training dataset obtained using an out-of-bag estimate.
+
+    `oob_prediction_` : array, shape = [n_samples]
+        Prediction computed with out-of-bag estimate on the training set.
+
+
 
     Notes
     -----
@@ -562,6 +626,7 @@ class RandomForestRegressor(ForestRegressor):
                        max_features="auto",
                        bootstrap=True,
                        compute_importances=False,
+                       oob_score=False,
                        n_jobs=1,
                        random_state=None):
         super(RandomForestRegressor, self).__init__(
@@ -571,6 +636,7 @@ class RandomForestRegressor(ForestRegressor):
                               "min_density", "max_features", "random_state"),
             bootstrap=bootstrap,
             compute_importances=compute_importances,
+            oob_score=oob_score,
             n_jobs=n_jobs,
             random_state=random_state)
 
@@ -638,6 +704,10 @@ class ExtraTreesClassifier(ForestClassifier):
         Whether feature importances are computed and stored into the
         ``feature_importances_`` attribute when calling fit.
 
+    oob_score : bool
+        Whether to use out-of-bag samples to estimate
+        the generalization error.
+
     n_jobs : integer, optional (default=1)
         The number of jobs to run in parallel. If -1, then the number of jobs
         is set to the number of cores.
@@ -652,6 +722,12 @@ class ExtraTreesClassifier(ForestClassifier):
     ----------
     `feature_importances_` : array of shape = [n_features]
         The feature mportances (the higher, the more important the feature).
+
+    `oob_score_` : float
+        Score of the training dataset obtained using an out-of-bag estimate.
+
+    `oob_decision_function_` : array, shape = [n_samples, n_classes]
+        Decision function computed with out-of-bag estimate on the training set.
 
     Notes
     -----
@@ -672,6 +748,7 @@ class ExtraTreesClassifier(ForestClassifier):
                        max_features="auto",
                        bootstrap=False,
                        compute_importances=False,
+                       oob_score=False,
                        n_jobs=1,
                        random_state=None):
         super(ExtraTreesClassifier, self).__init__(
@@ -681,6 +758,7 @@ class ExtraTreesClassifier(ForestClassifier):
                               "min_density", "max_features", "random_state"),
             bootstrap=bootstrap,
             compute_importances=compute_importances,
+            oob_score=oob_score,
             n_jobs=n_jobs,
             random_state=random_state)
 
@@ -749,6 +827,10 @@ class ExtraTreesRegressor(ForestRegressor):
         Whether feature importances are computed and stored into the
         ``feature_importances_`` attribute when calling fit.
 
+    oob_score : bool
+        Whether to use out-of-bag samples to estimate
+        the generalization error.
+
     n_jobs : integer, optional (default=1)
         The number of jobs to run in parallel. If -1, then the number of jobs
         is set to the number of cores.
@@ -763,6 +845,12 @@ class ExtraTreesRegressor(ForestRegressor):
     ----------
     `feature_importances_` : array of shape = [n_features]
         The feature mportances (the higher, the more important the feature).
+
+    `oob_score_` : float
+        Score of the training dataset obtained using an out-of-bag estimate.
+
+    `oob_prediction_` : array, shape = [n_samples]
+        Prediction computed with out-of-bag estimate on the training set.
 
     Notes
     -----
@@ -783,6 +871,7 @@ class ExtraTreesRegressor(ForestRegressor):
                        max_features="auto",
                        bootstrap=False,
                        compute_importances=False,
+                       oob_score=False,
                        n_jobs=1,
                        random_state=None):
         super(ExtraTreesRegressor, self).__init__(
@@ -792,6 +881,7 @@ class ExtraTreesRegressor(ForestRegressor):
                               "min_density", "max_features", "random_state"),
             bootstrap=bootstrap,
             compute_importances=compute_importances,
+            oob_score=oob_score,
             n_jobs=n_jobs,
             random_state=random_state)