+++ sklearn/metrics/cluster/supervised.py
@@ -498,7 +498,7 @@ def v_measure_score(labels_true, labels_pred):
 
 
 def mutual_info_score(labels_true, labels_pred, contingency=None):
-    """Adjusted Mutual Information between two clusterings
+    """Mutual Information between two clusterings
 
     The Mutual Information is a measure of the similarity between two labels
     of the same data. Where P(i) is the probability of a random sample occuring